move test functions to a separate file
