+++ symengine/tests/test_arit.py
@@ -1,4 +1,4 @@
-from symengine.utilities import raises
+from symengine.test_utilities import raises
 
 from symengine import (Symbol, Integer, Add, Mul, Pow, Rational, sqrt,
     symbols, S, I, count_ops)
