+++ symengine/tests/test_lambdify.py
@@ -8,10 +8,9 @@ from functools import reduce
 import itertools
 from operator import mul
 import math
-import sys
 
 import symengine as se
-from symengine.utilities import raises
+from symengine.test_utilities import raises
 from symengine import have_numpy
 import unittest
 from unittest.case import SkipTest
