+++ symengine/tests/test_number.py
@@ -1,4 +1,4 @@
-from symengine.utilities import raises
+from symengine.test_utilities import raises
 
 from symengine import Integer, I, S, Symbol, pi, Rational
 from symengine.lib.symengine_wrapper import (perfect_power, is_square, integer_nthroot)
