+++ symengine/tests/test_printing.py
@@ -1,5 +1,5 @@
 from symengine import (ccode, unicode, Symbol, sqrt, Pow, Max, sin, Integer, MutableDenseMatrix)
-from symengine.utilities import raises
+from symengine.test_utilities import raises
 from symengine.printing import CCodePrinter, init_printing
 
 def test_ccode():
