+++ symengine/tests/test_solve.py
@@ -1,4 +1,4 @@
-from symengine.utilities import raises
+from symengine.test_utilities import raises
 from symengine import (Interval, EmptySet, FiniteSet, I, oo, Eq, Symbol,
     linsolve)
 from symengine.lib.symengine_wrapper import solve
